Move image_map into yaml_frame_reader.rs
yaml_frame_reader is the only reader that supports external images.
The font map is also already in yaml_frame_reader.
